List Style fix(issue #50)
Not the most elegant fix, but it works. The problem was, it the php
function nl2br - adds <br /> with lists. We do not want this, as we can
format the text (new lines) when the user inputs it. The only other way
is creating a new nl2br like function, which could be complicated.